qbman: portable cacheline use
This uses cacheline manipulation wrappers that generalise to 32-byte
cachelines. It also refactors the portal register accessors to use the
__raw_[read|write][bl]() interfaces. The reason is that this code is
common/single-source with the kernel, but the kernel won't accept patches
that use "volatile" (because "it's nearly always wrong"). In our case
however, volatile access is exactly what we need. It happens that the
__raw_*() accessors implement precisely the volatile access we were using,
so deferring to those means our driver (and patches) don't violate any
coding standards or checker scripts.

This change also marks all pointers to read-only ring entries as const.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>