compat/conf: expose symbols that APIs use
Most of these symbols are required by driver APIs or PPAC APIs. As we
shouldn't require external code to access include/internal/*.h headers,
move the required definitions to the relevant include/usdpaa/*.h headers.

Note, some of these are needed by patches that are yet to come, but this
helps keep the patches disentangled.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>