qbman: merge portal-sharing from the kernel
The main portal drivers, including all single-source files, now support
the portal-initialisation flags that enable portal sharing across CPUs and
threads. However there is no USDPAA [qb]man_driver.c support to initialise
portals with these options enabled as yet. Whereas the kernel has an
up-front driver initialisation which determines in advance which CPUs
should get which portals (and which portals, if required, should be set
up for sharing, and for use by which CPUs), the USDPAA has no such scheme.
Instead, threads request initialisaton on an ad-hoc basis with portals
being allocated out on a first-come, first-serve basis, so the
USDPAA-specific APIs would need to be adapted to allow applications to
make these "choices". So while portal sharing logic is now in place for
USDPAA, it can't yet be turned on or tested.

NB, at this time, this completes the reunification of kernel and USDPAA
Qman/Bman drivers.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>