ppac: add per-thread PPAM init hooks
PPAM already has global init/finish hooks, so these are per-thread
equivalents. Note, the code had to be reordered slightly to ensure these
were declared before they were used.

Also, the global init hook should return zero on success, so fix the
weakly-linked (ie. overridable) default to do precisely that.

It is desirable for the thread-init hook to occur after global
initialisation, which required a little rewiring in the thread setup code.
(Normally, global init is performed in response to a message posted to the
primary thread and handled in its main loop, which would make it too late
for the thread-init hook. So instead we start the primary thread with its
message struct preset to the global-init message so that this is handled
prior to the app loop and the PPAM thread-init hook.)

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>