ppac: support a polling hook for PPAM
A global, thread-local boolean is used to control whether the per-thread
application-loop calls out to a PPAM hook "ppam_thread_poll()" or not.
PPAM can toggle this value on and off as required, either during init,
while handling network events, and/or from within the polling hook itself.
The weakly-linked default in PPAC will abort() if executed, implying that
PPAM should never toggle the boolean on unless it has also implemented
the hook function.

Note, the reason for the toggle instead of just calling the hook and
letting it decide whether to do anything or not is that the hook itself
can't be optimised via inlining (unless all of PPAC moves into the inlined
headers by dependency), so this is the best-performance alternative.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>