Fixed PPAC linker script so that USDPAA apps can start with valgrind.
The older linker script creates a read-write section named .rodata.cli_table.
This section is in middle of other read-only sections. The final executable
mapped these read-only and read-write sections with common segment.
This creates problem with valgrind.

http://old.nabble.com/-Bug-191290--New%3A-Assertion-%27rw_svma_limit-!%3D-0%27-failed-td23340625.html

To fix the issue, CLI commands are now added in a read-only section.

Signed-off-by: Vakul Garg <vakul@freescale.com>