fsl_fman: cleanup CRC64 API and consolidate headers.
This merges the two headers from include/fsl_fman/ (crc64.h and
fsl_fman.h) into a single header; include/fsl_fman.h.

This also cleans up a number of things in the CRC64 implementation;
 * crc64_*() APIs are renamed to fman_crc64_*()
 * CRC64_* symbols were either unused or were only used in one place, so
   their declarations have been removed as their presence was confusing.
 * comments have been simplified and fixed for spelling and grammar.
 * it turns out that the crc64_compute_<32|16|8>bit() functions were
   really treating their arguments as byte arrays rather than numerical
   values (ie. they're assumed to be in network byte order, irrespective
   of what the host byte order is). The algorithm also works on a
   byte-by-byte manner, so a single "compute" function is used to
   implement an arbitrary-width checksum ("fman_crc64_update()"), and the
   existing APIs are rewritten as wrappers around it. This makes the code
   clearer and shorter.
 * the global CRC64 table is no longer instantiated as a static global in
   every C file that includes it, but it is "extern"ed in order to link
   against a single instance. The Fman driver instantiates the table, but
   (as documented in the header), an application can instantiate it using
   a macro if it does not intend to link with the Fman driver.
 * the global CRC64 table (macro) is declared with 2 64-bit values per
   line of code rather than 1, taking it from 256 lines of header down to
   128.
 * some spurious white-space issues were cleaned up along the way.

This patch also updates the CRC64-dependent ipfwd and ipsecfwd code to
compile with these changes. It seems that numeric-typed variables are
being passed around in that code ("saddr", "daddr", "spi", and a casting
of "key" to uint32_t) in a way that will only work if the host is
big-endian *or* if the software is implicitly treating these values as
opaque groupings of bytes in network byte order. No attempt was made to
audit this code nor comment it for this purpose though.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>