of: use scandir(versionsort) instead of readdir()
Any ordering of nodes within the device-tree is lost in the file-system
representation, the best we can do is enforce our own ordering. The use of
the versionsort() callback in scandir() gives us a more intelligent
ordering than the usual alphasort(), particularly as they apply to the
naming of device-tree nodes. Eg. it will put "foo9" before "foo10".

This also tidies up some of the associated code, and adds a comment about
a leak in the error-handling (it was already there and this code doesn't
fix that). Another change is that any allocation or file-system errors in
the recursive parsing of the device-tree will now percolate to the
top-level of_init() call, which would previously only return failure if it
failed to open "/proc/device-tree".

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>