qman: remove CONFIG_FSL_QMAN_NULL_FQ_DEMUX
The current implementation of this feature is complicating other
development work, in part because it has not been done "quite right".
DC_ERNs are the only potentially useful aspect of this feature so far
and it isn't implemented correctly (this will be corrected in a
separate patch). Also ERN handling is not part of NULL demuxing, yet
the use of the "struct qman_fq_cb" type implies a default ERN handler.
Furthermore, for these kind of global handlers to be usable, it should
be possible to set them globally (ie. for use by all portals managed
by the driver instance) which the current implementation doesn't
allow.

This feature will be reimplemented properly once current rework is
done.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>