qbman: always compile portal support
There is little usefulness in compiling qbman drivers without support
for portals, and supporting such configurations complicates efforts to
reorganise the portal initialisation code to support for "dynamic"
use-cases. This removes the FSL_BMAN_PORTAL and FSL_QMAN_PORTAL
options and replaces the corresponding #ifdef-ery to the effect that
these options are permanently enabled.

For similar reasons of simplication, the now-pointless support for
disabling DCA support is removed. This harks back to a time when the
software was being run against simulation models that didn't support
DCA.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>