qbman: remove "legacy" FQID/BPID allocation, update dts
Both FQID allocation and BPID allocation are converging on a new means
for managing resources, and it is necessary to remove support for the
legacy mechanisms in order to complete the transition. The qm_fq_*()
APIs for FQ allocation are removed, as they no longer are needed to
cater to the legacy "bpid0" scheme (without which, these APIs were
simply front-ends to qman_[alloc|release]_fqid*() APIs). This commit
syncs with the same change on the kernel side, but is not a big change
because usdpaa wasn't using the old allocation schemes in the first
place.

This also factors out FQID/BPID allocation code into a common source
file, dpa_alloc.c, replacing qman_fqalloc.c. Note also that BPID
allocation was previously only possible by creating a pool object with
BMAN_POOL_FLAG_DYNAMIC_BPID. Now it is possible to explicitly allocate
BPIDs in the same way as with FQIDs, ie. in arbitrary-sized ranges
and/at arbitrary "alignments" if required.

Also, fix the API documentation - the "release" functions are
void-valued, so don't "return zero for success".

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>