qbman: use dynamic portal->cpu affinity
Portals are no longer statically bound to CPUs, in the device-tree nor
otherwise. So rather than a 'cpu' parameter being used to find a matching
portal within [qb]man_thread_init(), it instead just tries opening any
available portals until one opens successfully, at which point the kernel
backend will automatically bind it to the cpu the thread is executing on.

A side-effect of this is that a thread must now be affine to a single
core *before* calling [qb]man_thread_init(), so the usdpaa drivers check
that this is the case and will fail the API if it is not.

The equivalent change on the kernel side has also removed the linkage
between portals and pool-channels, for the same reason that portals are
now interchangable so distinctions between them in this regard are no
longer meaningful (such distinctions were only being (mis)used for
software configuration).

Note, the addition of 'liodns[2]' to qm_portal_config is due to sharing
source with the kernel drivers, where the PAMU manipulations are actually
requested. This field is unused in the USDPAA drivers.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>