Revert "of: parse virt->phys cpu mappings"
The need to be able to convert the virtual/contiguous cpu numbering of
user-space/pthreads to the physical/potentially-discontiguous cpu
numbering used in the device-tree, namely for making portals affine to
such a cpu, has gone now that the kernel makes portals affine on our
behalf. As such, we no longer need this cpu-mapping fluff and should
actively discourage its (mis)use by removing it altogether.

This reverts commit 0e5e6737eebd42127c9e065e41c4e2d946fe391e.