process: former dma_mem device will be multi-function
The file-descriptor underlying dma_mem was renamed from fsl-usdpaa-shmem
to fsl-usdpaa recently because it will be used for more than just mapping
DMA memory, it will be used for any other process-wide (and
non-portal-specific) interactions, such as FQID/BPID allocations. So it's
now a "USDPAA process driver", hence the "process" naming.

This splits the device-handling out into the "process" driver, leaving
dma_mem as a user of it (via internal-only interfaces, the intention is
for "process" to provide no externally-visible interfaces, but to provide
the kernel interactions underneath drivers that do expose APIs).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>