process/qman: add pool-channel/CGR allocators.
This syncs the new dynamic allocator APIs for pool-channels and CGR IDs
from the kernel, but in the USDPAA case these (de)allocations are proxied
through the process driver to the kernel-side allocators.

This also removes the placement of hard-coded ranges of pool-channels and
CGR IDs in the usdpaa_netcfg config data-structure, and instead hooks up
the process and qbman drivers to use kernel-side dynamic allocation of
these resources. The PPAC code has been adjusted to allocate CGRs (if the
feature is enabled) and pool-channels, as has hello-reflector.

The "fra" application has temporarily been disabled from the build, as it
needs adaptation in order compile.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>