dma_mem: update dma_mem for multi-process support
This overhauls the dma_mem driver to support the new kernel interface (ie.
this patch is interdependent with the corresponding kernel patch
"fsl_usdpaa: extend interface for multi-process support"). With this
change, there is no more assumption of there being a single DMA-enabled
memory mapping for each USDPAA application. Each process can create new
DMA maps and/or map to existing DMA regions created in other processes.
These maps may optionally have an allocator overlaid upon them (which is
synchronised across processes in the case of non-private regions), or they
can be left "raw" for applications to manage them as they see fit.

Applications and drivers that wish to assume the availability of a default
DMA region can use the global variable "dma_mem_generic". The assumption
is that the application would create and set this map, and legacy-style
APIs (called "__dma_mem_*") will use it. In particular the updates to the
pme2 and sec drivers rely on this map being set by the application, and
the updated applications are modified to create a process-private DMA
region and set it as the generic map.

Also, fix error-path handling in lpm-ipfwd - free() can not be used to
release memory alloccated from dma_mem_memalign().

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>