hello_reflector: don't use hard-coded BPIDs
This new buffer-pool-seeding behaviour is to have a default triplet of
0:0:1728, which can be overriden by the command line argument "-b
<x:y:z>", that is applied to each network interface's set of up-to-three
buffer pools. Where multiple network interfaces refer to the same buffer
pool(s), only the first occurance will be used. (So if bpid 7 is the
first pool mentioned in the first interface and the third pool mentioned
in some other interface, the pool will only be seeded according to the
first value in the triple - even if this is zero.)

Note, this is required for running multi-process.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>