build: fix insufficiencies of "make clean" target
The original "make clean" was only cleaning the object files that formed
the bins and libs, it wasn't removing the bins and libs as well. This
change harnesses the recursive processing of Makefile.am files to build
up a list of all the build-generated files, meaning the "clean" target
simply removes everything on that list.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>