build: don't automatically "touch" Makefile.am
The build was using Makefile.am files as dependencies that would be
touched as required when updating. All compiled objects were dependent on
their Makefile.am file and those in all parent directories up to the root,
and as such the arbitrary date-stamps of the Makefile.am files in a tree
would cause needless "touch"ing of other Makefile.am files. One
side-effect of this is that some git tools (eg. gitk) think that lots of
Makefiles are locally modified, but with zero-length diffs. (Running
"git status" rectifies this.)

This changes the build to instead use shadow files (".Makefile.touch") for
each Makefile.am, which are dependent on their corresponding Makefile.am
as well as reflecting the desired hierarchical dependencies. The creation,
touching, and deleting of these files is less annoying because they don't
contain code and are hidden from git - it avoids the need to update the
modification times of source files when haven't actually been modified at
all.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>