qbman: sync with kernel source
The updates to [qb]man_private.h have no functional impact on USDPAA, as
the changed prototypes are not implemented in user-space.

The other updates replace "normal spinlocks" with "raw spinlocks" in the
portal-sharing code. As portal-sharing is not compiled in USDPAA,
there's no need (yet) to add support for them in compat.h.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>