ppac: move the fd_set and tv variable initialization outside the loop
There are two reasons to do this modification.
1. If declare and initialize the fd_set and tv variables in the loop,
these variables will be created end released each loop, so, this will
impact the performance.
2. Usdpaa application sometimes is blocked by select function a few seconds.
It seems the tv's value has been changed to 10s, although in fact it is
always 10ms. After move the tv initialization outside the loop, the select
function will do correctly, and the program can response command immediately.
This patch is only a workaround for this issue.

This patch also changes fq_qman fd_bman nfds declaration and initialization,
to avoid the compile warning when not define FRA_IDLE_IRQ or PPAC_IDLE_IRQ.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>