fra: a workaround for RMan errata A-003585
Description: The command/status field of inbound Frame Descriptor is
overlapped with Frame Managers command bits. When the consumer of
inbound Frame Descriptor is another HW device there must be a common
format for the command/status fields to be redefined.

Workaround: Do not pass the Frame Descriptor from RMAN to FMAN directly.
Instead, use RMAN to pass the Frame Descriptor to e500 core and have
software to override the FD command/status field and enqueue back to FMAN.

This patch creates a copy of the original inbound frame descriptor, and
clears status field to avoid impact on FMAN.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>