fra: remove the outbound fd's RapidIO type field setting
Bits 96-99 of the outbound frame descriptor is optional by the
producer and is returned as-is during completion. The original
code used this field to describe RapidIO type, but it used a
wrong way to directly cast away "const" attribute of the fd. This
patch removes this part of the code and does not affect other
functions.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>