raw_test: simple sanity-test of the "raw" DMA map.
"raw" means no allocator is implemented on top of the DMA region, ie.
that the DMA_MAP_FLAG_ALLOC is not specified when creating the mapping.
In this case the user calls dma_mem_raw() to access the memory region
directly.

This is copied from dma_test, which always has the ALLOC flag set and
which has a useful CLI for doing test allocations and deallocations,
whereas raw_test has neither.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>