fman: add memac support
T4/B4 use fsl,fman-memac for both 1G and 10G ports, so we need to use
phy-connection-type to differ the mac type. Also memac uses command_config
register to enable Rx and Tx, which has the same offset as 10g GEC, so this
patch allows the correct setting to enable Rx/Tx.

Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>