Enable External Buffer Deallocation for FManv3
The new FMan v3 on T4/B4 needs to set EBD bit in contextA to allow buffer
deallocation because the FMD code does not set FMBM_TFNE[EBD], it also needs
to set FCO according to the "Tx confirmation enqueue and buffer deallocation
decision" chart in FMAN V3 RefManual to make EBD setting in contextA override
the EBD setting in FMBM_TFNE.

Reflector and hello_reflector have been tested with this patch on both B4860QDS
and T4240QDS. However, the changes for ipfwd, lpm-ipfwc and ipsecfwd are just
made to pass the compilation. There might be changes needed for those
applications.

Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>