fra: modify fra configuration analytical processing
The each separate configuration file contains all the configuration. There
are many repeated configuration This is not conducive to maintaining. We
derive the common configuration to a separate file as the default
configuration file. A new configuration file can include these files and joins
its specific configuration to modify the default value. This patch is to
add this functionality.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>