dpa_offload: Add DPA statistics offloading component
The DPA Stats library is the user space control component of the
DPA Stats Driver. It allows user space applications to access
DPA Stats services.
The purpose of the DPA Stats module is to provide to the application
a unitary method for retrieving counters that are spread at different
hardware or software locations.

Signed-off-by: Anca Jeanina FLOAREA <anca.floarea@freescale.com>
Change-Id: I1b4e0bc782c00b0d2e9317ccd7ed3b05cd233d72
Reviewed-on: http://git.am.freescale.net:8181/489
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>