sec: remove unused struct members
These members were meant to be populated by dcl but now they serve no
purpose other than to hold the encap/decap descriptors. The descriptors
are created with no relation to this structure which can just be
replaced by arrays of the propper size.

The array size is set to maximum to avoid buffer overflows (their
size can be reduced if memory is a constraint).

With this patch the wasted space is increased from ~20 bytes to ~160
bytes/tunnel in average. This figure is calculated by assuming an wasted
space of about 40 words/descriptor. In reality, the descriptors are
larger (34/27 words) so the wasted space is smaller.

Change-Id: Id560d14eb598c04aa918171898d940a9b78d8274
Signed-off-by: Cristian Stoica <cristian.stoica@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/1205
Reviewed-by: Porosanu Alexandru-B06830 <alexandru.porosanu@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>