dpa_offload: Add DPA packet classification offloading component
The packet classification offloading driver implements its functionalities
using the NetCommSw FMan driver. It exposes an easy-to-use API which can be
called either from kernel space or from user space (through the
classification offloading driver wrapper).

It is able to create or import (from FMD resources) 3 types of tables -
exact match table, indexed table and HASH table. Imported tables can be
either empty or prefilled.

It offers an API to insert, remove, or modify entries. It allows the users
to classify and
     - enqueue
     - multicast
     - discard
     - re-classify or
     - return-to-KeyGen
network packets.

It is able to create or import (from FMD resources) header manipulation
operations and attach them to table entries. It allows runtime modification
of existing (created or imported) header manipulation operations.

It offers an API to create or import (from FMD resources) multicast groups.
It allows the user to add or remove members to existing multicast groups.

Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/497
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>
Change-Id: I1fe382626d531533a80bffe07f69415172e5adb2
Reviewed-on: http://git.am.freescale.net:8181/2246
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>