dpa_offload apps: Add classifier_demo
The classifier_demo is a demo application which demonstrates some of
the capabilities of the packet classification offloading driver. It
configures a simple PCD which receives packets on one of the Ethernet
ports, classifies them into a variety of queues according to a specific
criteria and then sends the frames back to the Ethernet port they came
from.

The application also configures the FMan to offload header manipulation
for each of the classified flows.

Change-Id: If1bcb8f0935e3dbc776073610417df42e28049c8
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/496
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>