dpa_offload apps: Add fragmentation_demo
The fragmentation_demo is a basic test of offloading IP fragmentation
to the FMan. The application configures a PCD which uses one Rx/Tx port
and one offline port (which is doing the fragmentation). Both IP
fragmentation of IPv4 and IPv6 packets is supported.

The network packets received on the Rx port are classified according
to IP version and the protocol headers are updated. The frames are then
sent to the offline port where they are fragmented according to a
predefined MTU. The resulted fragments are returned to the port that
the original packets were received on.

Change-Id: Ifbc61d14bac3adc0b3e62b3977783dd2f2989947
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/499
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Wang Haiying-R54964 <Haiying.Wang@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>