dpa_offload apps: Add reassembly_demo
This test applications shows the FMAN IP Reassembly feature usage
and functionality. The IP Reassembly functionality can be applied on a
RX port or on an Offline port and performs reassembly on detected IPv4
and IPv6 protocol fragments. Also the DPA Stats component usage is
highlighted, by creating single counters for Ethernet and Reassembly and
both single and class counters for Classification Nodes.

For B4860QDS platform and other FMANv3 capable platforms the test
application has support for Virtual Storage Profiles. Three VSP's are
defined in the application and a set of buffer pools are allocated to
each VSP. This mechanism is totally transparent to the user.

Traffic in the form of IP fragments is received on a RX port by
reassembly_demo application. Then the fragments are reassembled, by the
FMAN ucode, the Ethernet MAC addresses are swaped and enqueued in the TX
queue of the same port by the test application.

The user has the possibility to read the values of the DPA Stats
counters by typing in the USDPAA PPAM command line: get_stats (retrieve
statistics for the counters in asynchronous mode), get_stats_sync (for
syncronous mode) and reset_stats (for reseting statistics)

Change-Id: I37923dd3f85f64d089715a446d1f6adf269495ca
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/409
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Wang Haiying-R54964 <Haiying.Wang@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>