simple_proto: fix MACsec for SEC ERA 2
MACsec protocol fails when running on SEC ERA 2 (i.e. P4080 TO2).
Apparently the DECO register used for saving the PN from PDB (M0)
is used by the MACsec OPERATION PROTOCOL command. Also, the shared
descriptor has to wait for the actual PN to be written in physical
memory. This is needed because it's possible that another DECO fetches
the descriptor/PDB from the physical memory (which could be
already modified by the previous executing MACsec OPERATION PROTOCOL
command).

Change-Id: I2a020acc88f57edc1cf5b13ced0c2b999f614ca3
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2299
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>