simple_proto: fix buffer overflow
The "protocol" variable, which holds the name of the protocol
to be tested,  could accommodate strings of up to 19 characters
in length. For PDCP protocol processing, the length of the
protocol name contains other information as well, making the
total length > 19 characters. The bug was manifesting itself
only when using the compiler in Yocto from SDK 1.4.

Change-Id: I92f5accd9994a32439d400d3e13fd9c7ac2c1c87
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2708
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>