simple_proto: fix WiMAX frames checking
- test that only the first output encapsulation/decapsulation frame
matches with the corresponding output encapsulation/decapsulation
test vector for WiMAX in CIPHER mode.
SEC increments and writes back the Packet Number (PN) to the PDB
for each frame processed. When WiMAX application testing is done
by inserting the same test vector repeatedly, we will obtain
output frames that differ due to the PN incrementation.

- set the input plain-text data as GMH aware frames
for WiMAX in PERF mode. The GMH Header Type bit shall be set to zero,
the CRC indicator bit shall be set to one (if CRC is included in the PDU
by appending it to the PDU payload encryption), and the GMH LEN field
shall be updated accordingly to the input plain-text data length,
without SEC intervention.

- fix the length by which the frames are verified in
test_dec_match routine for WiMAX PERF mode.

Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Change-Id: I8a61fa0906596d765bb6a5c891ea34ef93b186ef
Reviewed-on: http://git.am.freescale.net:8181/2766
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>