standalone.env: enable out-of-tree & host arch based compilation
This patch adds the necessary magic to enable the user to compile
USDPAA when the sources are not under the SDK Devel hierarchy.
Also, it takes into consideration both the host architecture and
target architecture for selecting the proper compiler path.
This patch also unifies the standalone-env and standalone-env.powerpc64
files (i.e. adds support for 64 bit compilations in USDPAA in
standalone-env), thus standalone-env.powerpc64 is no longer
required.

Change-Id: I3a5df02a479aa158165f0d5e88cbcf9abe9697cd
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2862
Reviewed-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>