reflector : Block the removal of the primary thread on lookup via uid
Current code disallows the removal of the primary thread on lookup
via cpu. But it doesnt give protection against removal via uid.
When the user removes the primary thread via uid and issues quit
command, the application fails to exit gracefully as the attempt
is being made to access the already freed memory. The user needs to
issue ctrl+C to exit the app. Hence add the check to block the
removal of the primary thread via uid also.

Signed-off-by: Kuldip Giroh <kuldip.giroh@freescale.com>
Change-Id: Ic596bbae6bce895decddf8c4216cdedc629961ae
Reviewed-on: http://git.am.freescale.net:8181/3060
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>