dpa_offload: Add support for header manipulation on DPA Classifier next_table action
The DPA Classifier API did not allow the user to add header
manipulation on an entry whose action was sending the frames to
a new classification table. This change adds this support
to the DPA Classifier by adding an "hmd" (header manipulation
descriptor) attribute to the "dpa_cls_tbl_next_table_desc"
parameters structure.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ifdd084df0c094df481bfde36530e7fe1e0c40a56
Reviewed-on: http://git.am.freescale.net:8181/3105
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>