hello_reflector : Add support to run the app in short circuit mode
User may want to run the application in short circuit mode.
Here packets are dequeued from the interface and received on
Rx FQs. These packets are then sent out from the QMAN channel
on which Tx FQs are scheduled. We can be sure of the sanity
of the hardware path for the packet flow by this way without
any processing by the core on the received packets.

Signed-off-by: Kuldip Giroh <kuldip.giroh@freescale.com>
Change-Id: I0701d54bd4597167e7793ee8b7bc75cf00997503
Reviewed-on: http://git.am.freescale.net:8181/3236
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>