dpa_offload: Add separate resource pointer for multicast group import operation
When multicast group was preallocated from a xml PCD file, the handle to the group
was given as a multicast group parameter.
To align the multicast component with header manipulation and classifier
components, the group handle will be given as a resource input parameter
to the function that creates a multicast group.

Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Change-Id: If79f8a7af4def15854b661f0503c2e634d4d5577
Reviewed-on: http://git.am.freescale.net:8181/3280
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>