reassembly_demo: Remove frame queue used for IP reassembly timeout
The frame queue used for incomplete frames that timed out during IP
reassembly was added as a workaround to avoid a known issue in the
FMan ucode related to releasing buffers from reassembled frames. This
workaround is not necessary anymore since ucode 106.x.8 and it needs
to be removed in order not to cause other problems. The frame queue
for timeout events is now specified as zero, which means that the
ucode will drop all incomplete frames that have timed out during IP
reassembly.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I60e2cb1e8bbc398abe15b33cc4110bd93b87cc33
Reviewed-on: http://git.am.freescale.net:8181/4181
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>