dpa_offload: Added Traffic Manager US counter implementation
Added support for Traffic Manager DPA Stats user-space counters. Until
now all DPA Stats counters were treated in kernel-space, but for QMan
CEETM counters, the mechanism for obtaining the statistics must be
implemented in user-space. For every user-space counter a thread that
has an affine portal mapped is created and the QMan get_statistics
function is called. In case of mixed request, the kernel will provide
the memory offset, and the library will write the counter statistics in
the storage area accordingly. For the kernel only counters request no
changes were made.

Change-Id: Ib3ad1557d3d3ac573a20409d60af5fb297c408f8
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4197
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>