dpa_offload: Reduce the size of the IPv4 header options field
According to RFC 791, the size of the IPv4 header (options
included) is represented on 4 bits (IHL), hence it can be an
absolute maximum of 15 * 4 = 60 bytes. The IPv4 options buffer
size field in the fsl_dpa_offload.h header file, in the
ipv4_header data structure, was declared as an "unsigned int" and
this was causing a static analysis warning in dpa_classifier.

The field was reduced to uint8_t which is more than enough to
store the size of the IPv4 frame options.

Change-Id: Ifa43d7b14dc5450e407d6b2928920dafd91c7c9b
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4265
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>