ipsecfwd: sync with the parameters of RTA IPsec descriptors
The RTA declarations of the IPsec descriptors functions changed.
A redundant parameter was removed, and a new one was added.

The redundant parameter was concerning the encapsulation descriptor;
the Optional Ip Header, if present, has to be part of the PDB.

The new parameter determines the size of address pointers.

Change-Id: Ia874c85875d47d24e40610f96afb1ad276d27313
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4325
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>