dpa_offload: Memory allocation fixes in DPA Stats
This commit replaces the sequential memory allocations of some internal
DPA Stats structures with allocation of bigger memory chunks. Apparently
there is a problem when an application allocates the same quantity of
memory a number of times.
The problem related to pthread_condition_wait is also fixed by this
change.

Change-Id: I2bd68ad1fb0715818afc171924411153f12904d8
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4393
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>