ipsec : Add extended sequence number support
IPsec can use an extended sequence number (ESN) optionally that is
authenticated but not transmitted. If an ESN is found in the PDB,
it is given to the authentication CHA in the last. The ESN
is incremented whenever the Seq Num rolls over.

For this purpose, user is provided two optional parameters. One is
-x, which is intended to tell if user wants to use ESN option or
not. Other is -v, which the user can configure with some starting
sequence number for the packets. Both the options are independent
of each other. User may provide these values to the app via the
control plane.

Signed-off-by: Kuldip Giroh <kuldip.giroh@freescale.com>
Change-Id: I528c005e2a83a08c56693f2a9402b35393f7b602
Reviewed-on: http://git.am.freescale.net:8181/4227
Reviewed-by: Dutta Yashpal-B05456 <yashpal.dutta@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>