simple_crypto, simple_proto: add support for SEC ERA retrieval
If the user does not specify an era to be used by RTA
in order to generate and validate the descriptors,
the application will try to read the SEC ERA from
the /proc/device-tree pseudo-filesystem using the OF
library (part of USDPAA). If neither the user provides
an era, nor the corresponding SEC ERA node exists in
the /proc/device-tree structure, then the default SEC
ERA is used (ERA 2).

Change-Id: I5c204b837201db28996f8bedca47d25346d1d079
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4939
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>