dpa_offload: Use only one list node structure to manage async statistics counters user space requests
There were 2 separate list nodes in the "dpa_stats_async_req" structure
("node" and "us_node") that were used when the objects were added to the
available requests pool and when the request was added to the user space
requests list respectively. As the async request object cannot be part of
the two lists in the same time, the same list node structure can be used
when working with both lists.

Signed-off-by: Anca Jeanina FLOAREA <anca.floarea@freescale.com>
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Iece63f8cf618bd61e43481f195e3cfbdfc2f30fd
Reviewed-on: http://git.am.freescale.net:8181/4593
Reviewed-by: Zanoschi Aurelian-B43522 <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>